Adds forest fire simulation version for python 3.2.5
Adds forest fire simulation `forestfiresim_325.py` to align with the class requirement of 3.2.5 compatiblity

This commit includes a modified version of the forest fire simulation, adapted for compatibility with Python 3.2.5.
The original `forestfiresim.py` is enhanced with the addition of the lake functionality, making it compatible with `forestfiresim_325.py`.
The lake feature acts as a firebreak to prevent flames from crossing.